Initial commit
This is just a collection of Vim scripts from different places (some
from #vim, some from r/vim, some mine).

If I see something cool, it'd be a shame to lose it for reference later
on, although it's not all the time I actually want to put these in my
~/.vim/plugin.

I usually kept such scripts in a vim-playground directory; this is just
me checking it into Git.