Ignore sandbox directory
I also resort to my vim-playground for testing random snippets on
different kinds of text files. Having a sandbox for this testing would
be nice, since the extraneous files wouldn't clutter my working tree.

If a script in the sandbox gets a nice enough implementation to warrant
saving, I can simply move it out to project root. Git will neatly keep a
list of these scripts (as untracked files in my git status), as opposed
to if I had *all* test files as untracked files so I'd have to scour to
find which ones are worth keeping.