Add synstack.vim
This is a file that's been in my sandbox for a bit. It enables a
<plug>(SynStack) map (which is in turn mapped to <C-p>, by default) that
tells you what syntax highlight groups apply to the word under the
cursor.

I've lost track of the original author, although I *have* seen a lot of
snippets achieving this floating around on the Web. Perhaps this exact
file is from the Vim wikia, or from one of the regulars in #vim on
freenode.